requirements.txt
Django==1.7.1
Markdown==2.5.1
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.16
django-filter==0.8
djangorestframework==2.4.4
kombu==3.0.23
pika==0.9.14
psycopg2==2.5.4
pytz==2014.9
virtualenv==1.11.6
static==0.4
dj-database-url==0.3.0
dj-static==0.0.6
argparse==1.2.1
SQLAlchemy==0.9.8
django-sse==0.4.1
sse==1.2